Separate JavaFX components into own package
