+++ src/main/java/kronos/ui/components/DialogBox.java
@@ -1,4 +1,4 @@
-package kronos;
+package kronos.ui.components;
 
 import java.io.IOException;
 import java.util.Collections;
